feat(client): Invitation Card
