+++ packages/components/src/reserves/index.ts
@@ -1,4 +1,4 @@
+export * from "./InvitationSection";
 export * from "./PlayerInput";
 export * from "./ReservationInfo";
 export * from "./SelectionSection";
-export * from "./InvitationSection";
